+++ src/settings/MappingProfiles/initialDetails/MARC_BIBLIOGRAPHIC.js
@@ -1,10 +1,6 @@
 const MARC_BIBLIOGRAPHIC = {
   name: 'marcBib',
   recordType: 'MARC_BIBLIOGRAPHIC',
-  marcMappingDetails: [{
-    order: 0,
-    field: { subfields: [{}] },
-  }],
 };
 
 export default MARC_BIBLIOGRAPHIC;